requirements.txt
click
jsonschema
pip-check-reqs
pip-licenses
pylint
pytest
pytest-cov
pytest-pythonpath
pytest-pyspec
setuptools
Sphinx
tox
twine
yamllint
